Add the expat license. One can grab everything that was already CC0 and treat it that way, of course, but any substantial new stuff will be expat.
It is simpler IMO to treat the stuff as expat. That is why I am
switching to this. Public domain is harder to achieve and express than
an expat license. That is why I like the expat license.

(By contrast I deplore the Apache license, even though I have been
told it amounts to the same thing. I cannot even read it! It is
written in Venusio-Neptunian! For me, that is unacceptable.)